Correctly handle the conversion of POINT_WISE compositions
Fixes issue #14.

This commit also adds a warning when it detects an MCNP composition that
was specified with mass fractions (negative values in the Mn card), but
used with a total atom concentration in a cell (in atoms/cm³).  At the
moment we do not support this because it would require the use of an
atomic mass table (probably MCNP's?).