Handle fully-specified lattice cells with recursive universe inclusion
This commit adds support for the inclusion of the lattice universe in
one of the cells of a fully-specified lattice. What this does is that
the lattice cell is filled with the material of the base cell (the one
with LAT=1).

This completes support for fully-specified lattices and closes
issue #16.