Fix MCNP lexing
This commit fixes a bug in lexing/parsing that resulted in incorrect
conversion of cells that made liberal use of spaces. A side-effect of
this correction is that t4_geom_convert (actually MIP) now supports
TatSu >= 4.4.0.

This commit also adds a simple MCNP input test file to make sure there
are no regressions.